virt_mshv_vtl, snp: Read and report the SEV status just once
The SEV status in VTL2 is read via the MSR driver on each VP,
and is reported for each VTL. While it's certainly possible to
have different VMSA SEV features on different VPs, that's not
expected: that's treated as a partition-wide property.

Read the MSR once, cache that in the shared part of the backing,
and report once. That shaves off some time from starting up the
large VMs (was ~7 sec wall clock time for me for a 64VP VM on a
Turin host, serial debugging on).