openhcl: produce some diag data for the UEFI watchdog
Some stack frames are logged as well as the stack pointer and
the instruction pointer. Regretfully, this currently bypasses the existing
interfaces and doesn't take into account wthether the VM is a hardware
isolated VM or not. Shouldn't be used on the CVMs as is. Folks didn't
express concerns over leaking secrets so far (as running in UEFI).

Should be helpful in some investigations.