{underhill|vmm|hvlite}_core: Use the VS CPUID leaf to indicate Confidnetial VMBus is available.
The intended use is with the Confiddntial VMBus patches I've been
working on: https://lore.kernel.org/linux-hyperv/20250714221545.5615-1-romank@linux.microsoft.com/

In order to establish the Confidential VMBus connection, the conf.
guest first tries to connect to the paravisor using the non-CVM
specific mechanisms so that the intercepts go to the paravisor.
Then the guest performs version negotiation. If any of that fails,
the VMBus connection is not confidential, and the guest falls back
to using the GHCB/GHCI protocols when establishing the VMBus
connection.

The Linux VMBus driver may be compiled as a module and loaded from
the user mode. Given that it relies on the CPU hot-plug callbacks
to initialize the SynIC on each VP, and they being fallible leads
to failures in onlining CPUs, the fall-back sequence mentioned above
isn't the best fit for all sceanrios. An example would be the SynIC
implemented in the legacy HCL that doesn't support the Confidnetial
VMBus and thus leads to the issues in the mentioned edge case.

The agreed upon approach on x86 is to use a bit from the VS leaf.
While intended for use from the Linux VMBus driver, no harm if
any other implementation of the VMBus driver takes a dependency on
that on x86.

On ARM64 CC, there is no legacy HCL, and the Confidential VMBus
will be required so that the guests will not need to retry.