virt_mshv_vtl: 2505: work around the hypervisor's handling of the precise VP startup on arm64
This is a backport of 075cd7c3f84d3ed7923c999a32208b5d10af9c04
to release/2505

On arm64, the hypervisor may schedule APs in VTL0 for execution even
when the partition is created with the precise VP startup flag. That
leads to the AP starting executing code at `PC`=`0` in VTL0 and sooner
or later:
* running into an exception and jumping to the VBAR table, then jumping
around until the guest loads a good context on the AP,
* in rare cases the AP runs into a stage-2 page fault (such as jumping
out of its GPA) which leads to a VM exit to VTL2. As that's a fault to
fetch an instruction, VTL2 has no other choice other than crash.

Fix this by handling the startup suspend state in VTL2 on arm64.

That does not happen on x86, as handling of the startup suspend state
has been historically present in the paravisor

---------

Co-authored-by: John Starks <jostarks@microsoft.com>