minimal_rt: Rewrite memcpy and memset in Rust
The `memcpy` and `memset` intrinsics were impelemented in asm.
Assembly is harder to support and review, the SIMD extensions the
compiler may use are enabled before the Rust-compiled code is
run. The generated code is good enough for the bootloader case.
Another motivation would be the recent issue fixed in
https://github.com/microsoft/openvmm/commit/801e80887298726a1fdb9b75b32bdd287c42f329

Replace the assembly with idiomatic Rust to bring the source code to
a better form.